Restructure folders and update data
