+++ src/physical-and-chemical-data/interfaces/solubility.ts
