+++ src/physical-and-chemical-data/interfaces/utils.ts
